api/requirements.txt
black==22.1.0
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
Pillow==9.0.1
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-dotenv==0.20.0